refactor: organizing folder and file structure
