+++ packages/api/routes/city.ts
@@ -1,0 +1,4 @@
+import express from 'express';
+
+const router = express.Router();
+export const cityRouter = router;
