+++ packages/api/routes/country.ts
@@ -1,0 +1,4 @@
+import express from 'express';
+
+const router = express.Router();
+export const countryRouter = router;
