+++ packages/api/tsconfig.json
@@ -9,5 +9,16 @@
     "skipLibCheck": true,
     "baseUrl": ".",
     "outDir": "./dist",
+    "paths": {
+      "@routes/*": [
+        "routes/*"
+      ],
+      "@middlewares/*": [
+        "middlewares/*"
+      ],
+      "@db/*": [
+        "db/*"
+      ]
+    }
   }
 }
